Add component moved
